+++ src/novelystlib/view_controller/right_frame/chapter_view.py
@@ -25,18 +25,19 @@ class ChapterView(BasicView):
     _NOTES = False
     _BUTTONBAR = True
 
-    def __init__(self, ui):
+    def __init__(self, ui, parent):
         """Initialize the view once before element data is available.          
         
         Positional arguments:
             ui: NovelystTk -- Reference to the user interface.
+            parent -- Parent widget to display this widget.
 
         - Initialize element-specific tk entry data.
         - Place element-specific widgets in the element's info window.
         
         Extends the superclass constructor.
         """
-        super(). __init__(ui)
+        super(). __init__(ui, parent)
 
         # 'Do not auto-number...' checkbutton.
         self._noNumber = tk.BooleanVar()