+++ src/novelystlib/view_controller/right_frame/character_view.py
@@ -34,18 +34,19 @@ class CharacterView(WorldElementView):
     _LBL_X = 15
     # Width of left-placed labels.
 
-    def __init__(self, ui):
+    def __init__(self, ui, parent):
         """Initialize the view once before element data is available.
         
         Positional arguments:
             ui: NovelystTk -- Reference to the user interface.
+            parent -- Parent widget to display this widget.
 
         - Initialize element-specific tk entry data.
         - Place element-specific widgets in the element's info window.
         
         Extends the superclass constructor.
         """
-        super(). __init__(ui)
+        super(). __init__(ui, parent)
 
         #--- 'Full name' entry.
         self._fullName = MyStringVar()