+++ src/novelystlib/view_controller/right_frame/project_view.py
@@ -26,18 +26,19 @@ class ProjectView(BasicView):
     _NOTES = False
     _BUTTONBAR = False
 
-    def __init__(self, ui):
+    def __init__(self, ui, parent):
         """Initialize the view once before element date is available.
         
         Positional arguments:
             ui: NovelystTk -- Reference to the user interface.
+            parent -- Parent widget to display this widget.
 
         - Initialize element-specific tk entry data.
         - Place element-specific widgets in the element's info window.
         
         Extends the superclass constructor.
         """
-        super(). __init__(ui)
+        super(). __init__(ui, parent)
 
         #--- Author entry.
         self._authorName = MyStringVar()