+++ src/novelystlib/view_controller/right_frame/world_element_view.py
@@ -28,18 +28,19 @@ class WorldElementView(BasicView):
     _NOTES = False
     _BUTTONBAR = True
 
-    def __init__(self, ui):
+    def __init__(self, ui, parent):
         """Initialize the view once before element data is available.
         
         Positional arguments:
             ui: NovelystTk -- Reference to the user interface.
+            parent -- Parent widget to display this widget.
 
         - Initialize element-specific tk entry data.
         - Place element-specific widgets in the element's info window.
         
         Extends the superclass constructor.
         """
-        super(). __init__(ui)
+        super(). __init__(ui, parent)
 
         self._fullNameFrame = ttk.Frame(self._elementInfoWindow)
         self._fullNameFrame.pack(anchor=tk.W, fill=tk.X)